+++ src/main/java/org/pokeherb/orderservice/application/service/dto/request/OrderCreateRequestDto.java
@@ -1,17 +1,17 @@
-package org.pokeherb.orderservice.application.service.dto;
+package org.pokeherb.orderservice.application.service.dto.request;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-public record OrderCreateCommand(
+public record OrderCreateRequestDto(
         UUID productId,
         int quantity,
         UUID orderUserId,
         String productName,
         LocalDateTime dueAt,
         String requestMemo,
-        long startHubId,
-        long endHubId,
+        Long startHubId,
+        Long endHubId,
         UUID requestVendorId,
         UUID receiveVendorId
 ) { }
