+++ src/main/java/org/pokeherb/orderservice/presentation/controller/OrderController.java
@@ -1,11 +1,7 @@
 package org.pokeherb.orderservice.presentation.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.pokeherb.orderservice.domain.Order;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("v1/order")
